Avoid adding cache buster on remote commands
The CDN is now using sensible TTLs (60 minutes) so the cache buster is less relevant.
Also fix the implementation of that method, in case someone was using it.
Also deprecate it as we don't need it anymore for ourselves and will be removed in the future.
Some indentation fixes to remove swiftlint warnings.
Increase version to 2.18.2.